@@ -1,16 +1,17 @@
 plugins {
     id 'com.android.application'
     id 'com.google.gms.google-services'
+//    id 'com.google.gms.google-services'
 }
 
 android {
-    compileSdkVersion 31
+    compileSdkVersion 32
     buildToolsVersion "30.0.3"
 
     defaultConfig {
         applicationId "com.example.visage"
         minSdkVersion 21
-        targetSdkVersion 31
+        targetSdkVersion 32
         versionCode 1
         versionName "1.0"
 
@@ -38,23 +39,33 @@ dependencies {
     implementation 'com.google.android.material:material:1.5.0'
     implementation 'androidx.constraintlayout:constraintlayout:2.1.3'
     implementation 'androidx.legacy:legacy-support-v4:1.0.0'
-    implementation 'com.google.firebase:firebase-database:20.0.3'
+    implementation 'com.android.car.ui:car-ui-lib:2.0.0'
     implementation 'com.google.firebase:firebase-auth:21.0.1'
-    implementation 'com.android.car.ui:car-ui-lib:+'
+    implementation 'com.google.firebase:firebase-database:20.0.3'
     testImplementation 'junit:junit:4.13.2'
     androidTestImplementation 'androidx.test.ext:junit:1.1.3'
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.4.0'
 
+    //lottie and cuberto
     implementation 'com.airbnb.android:lottie:4.2.2'
     implementation 'com.cuberto:liquid-swipe:1.0.0'
     implementation 'androidx.core:core-ktx:1.7.0'
     implementation "org.jetbrains.kotlin:kotlin-stdlib:1.6.10"
+
+    //bottom navigation
     implementation 'com.etebarian:meow-bottom-navigation-java:1.2.0'
 
+    //login and signup
     implementation "androidx.recyclerview:recyclerview:1.2.1"
     implementation "androidx.cardview:cardview:1.0.0"
     implementation 'com.github.scottyab:showhidepasswordedittext:0.8'
+    
     implementation 'de.hdodenhof:circleimageview:3.1.0'
+
+    //map
     implementation 'com.google.android.gms:play-services-maps:18.0.2'
     implementation 'com.karumi:dexter:6.2.3'
+
+    //profile
+    implementation 'com.android.support:design:32.0.0'
 }
