@@ -9,9 +9,6 @@ import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.Button;
-import android.widget.EditText;
-
-import java.util.Objects;
 
 /**
  * A simple {@link Fragment} subclass.
@@ -45,6 +42,7 @@ public class HomeFragment extends Fragment {
      * @return A new instance of fragment HomeFragment.
      */
     // TODO: Rename and change types and number of parameters
+
     public static HomeFragment newInstance(String param1, String param2) {
         HomeFragment fragment = new HomeFragment();
         Bundle args = new Bundle();
@@ -63,7 +61,6 @@ public class HomeFragment extends Fragment {
         }
     }
 
-    private EditText location_finder;
 
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
@@ -73,15 +70,11 @@ public class HomeFragment extends Fragment {
         View view = inflater.inflate(R.layout.fragment_home, container, false);
 
         //String menu = getArguments().getString("Menu");
-        Button map = (Button) view.findViewById(R.id.map_test);
-        map.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
-                Intent i = new Intent(getContext(),Map_Activity.class);
-                startActivity(i);
-            }
-        });
-
+//        Button map = (Button) view.findViewById(R.id.map_test);
+//        map.setOnClickListener(view1 -> {
+//            Intent i = new Intent(getContext(),Map_Activity.class);
+//            startActivity(i);
+//        });
 
         return view;
 
